thread testing.T through the codebase
Delegate as much as possible of the API surface to the testing
package by embedding testing.T in check.C. In the process we
have to rip out the custom test runner, since it doesn't play
well with the assumptions made by the Go runner.

Functions which execute the testsuite now don't block anymore
waiting for tests to complete.